faceDetection/requirements.txt
python-dotenv
cvzone
ultralytics
mediapipe==0.10.21
numpy
opencv-python
scikit-learn
joblib
facenet-pytorch 
torch           
torchvision

flask
flask_sqlalchemy
SQLAlchemyError
flask_jwt_extended
psycopg2
werkzeug
flask-cors
flask_socketio
flask-migrate
flask-bcrypt
Flask-Mail

web3
py-solc-x